feat(release): Cycle 13 — Deck Evaluation & Creator Output (2026.06.10)
Squash-merge of branch plan/cycle-13-deck-eval (phases 75-78).

- Phase 75 Tap Analyzer Surface: untapped-source frequency + turn-1
  untapped availability off the existing castability simulation
  (flag analysis.manabase.tap-analyzer, OFF).
- Phase 76 Bracket Classifier + Balancer: official 5-tier classification
  + balancer paste artifact; Game Changers -> versioned bracket-data.json;
  own tool tile (flag tool.bracket.enabled, OFF).
- Phase 77 Multi-Axis Deck Score: Power/Speed/Control/Consistency 0-5
  bands in /deck-analysis + all three paste variants
  (flag analysis.multi-axis-score, OFF).
- Phase 78 Auto-Refreshing Primer: resume-without-rebuild stale banner
  on the Deck Primer; multiset-hash staleness, no auto-rebuild/re-fetch
  (flag tool.primer.stale-flag, OFF).

All four cycle flags seeded OFF. Build 0/0; CI green on the branch.

+++ DeckFlow.Web/Models/MetaGapResponse.cs
@@ -38,10 +38,11 @@ public sealed class MetaGapData
     public int RefDeckCount { get; init; }
 
     /// <summary>
-    /// Overall cEDH readiness score used to rank how close the deck is to the reference meta.
+    /// Overall cEDH readiness score (0-10) used to rank how close the deck is to the reference
+    /// meta. Stored as a double because AIs routinely return a fractional value (e.g. 9.7).
     /// </summary>
     [JsonPropertyName("readiness_score")]
-    public int ReadinessScore { get; init; }
+    public double ReadinessScore { get; init; }
 
     /// <summary>
     /// Explanation for the readiness score shown alongside the numeric rating.
